+++ facebook_auth/example/lib/pages/splash/widgets/loading.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class Loading extends StatelessWidget {
+  const Loading({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Positioned.fill(
+      child: Container(
+        color: Colors.white70,
+        child: Center(
+          child: CupertinoActivityIndicator(),
+        ),
+      ),
+    );
+  }
+}