+++ facebook_auth/lib/flutter_facebook_auth.dart
@@ -29,6 +29,24 @@ class FacebookAuth implements FacebookAuthPlatform {
   /// retrive the user information using the GraphAPI
   ///
   /// [fields] string of fields like birthday,email,hometown
+  ///
+  /// The facebook SDK will return a JSON like
+  /// ```
+  /// {
+  ///  "name": "Open Graph Test User",
+  ///  "email": "open_jygexjs_user@tfbnw.net",
+  ///  "picture": {
+  ///    "data": {
+  ///      "height": 126,
+  ///      "url": "https://scontent.fuio21-1.fna.fbcdn.net/v/t1.30497-1/s200x200/84628273_176159830277856_972693363922829312_n.jpg",
+  ///      "width": 200
+  ///    }
+  ///  },
+  ///  "id": "136742241592917"
+  ///}
+  ///```
+  ///
+  ///The above JSON could be change, it depends of your [fields] argument.
   @override
   Future<Map<String, dynamic>> getUserData({
     String fields = "name,email,picture.width(200)",
@@ -67,4 +85,10 @@ class FacebookAuth implements FacebookAuthPlatform {
       version: version,
     );
   }
+
+  /// returns one instance of FacebookPermission with the granted and declined permissions
+  ///
+  /// It could be null if you exceed the request limit
+  @override
+  Future<FacebookPermissions?> get permissions => _.permissions;
 }