+++ facebook_auth_platform_interface/lib/src/facebook_permissions.dart
@@ -1,0 +1,4 @@
+class FacebookPermissions {
+  final List<String> granted, declined;
+  FacebookPermissions({required this.granted, required this.declined});
+}